express-backend/requirements/base.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.2
psycopg2-binary>=2.9
dj-database-url>=2.2
python-dotenv>=1.0
celery>=5.3
redis>=5.0
stripe>=10.0
requests>=2.31
phonenumbers>=8.13
pyotp>=2.9
social-auth-app-django>=5.4
